requirements.txt
# Image Pipeline Requirements
# Tested with Python 3.10+

# Core ML frameworks
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
accelerate>=0.27.0

# Quantization for INT8 inference
bitsandbytes>=0.42.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Embeddings and vector store
qdrant-client>=1.7.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Model-specific dependencies
# SigLIP / CLIP
open-clip-torch>=2.24.0

# GroundingDINO
groundingdino-py>=0.4.0

# SAM2
segment-anything-2>=1.0

# Hugging Face model downloads
huggingface-hub>=0.20.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0

# Image hashing for deduplication
imagehash>=4.3.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# HTTP requests (for dataset downloads)
requests>=2.31.0
aiohttp>=3.9.0

# Utilities
rich>=13.0.0  # Pretty console output
typer>=0.9.0  # CLI interface
